Added toggle for watched and non watched movies

+++ app/routes/notes.$noteId.tsx
@@ -2,7 +2,7 @@ import type { ActionFunction, LoaderFunctionArgs } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
 import { Form, useLoaderData } from "@remix-run/react";
 import type { Note } from "~/models/note.server";
-import { deleteNote, getNote } from "~/models/note.server";
+import { deleteNote, getNote, toggleWatchedNote } from "~/models/note.server";
 import { requireUserId } from "~/session.server";
 import invariant from "tiny-invariant";
 
@@ -26,7 +26,17 @@ export const action: ActionFunction = async ({ request, params }) => {
   const userId = await requireUserId(request);
   invariant(params.noteId, "noteId not found");
 
-  await deleteNote({ userId, id: params.noteId });
+  const formData = new URLSearchParams(await request.text());
+
+  if (formData.get('action') === 'delete') {
+    await deleteNote({ userId, id: params.noteId });
+    return redirect("/notes");
+  }
+
+  if (formData.get('action') === 'toggleWatched') {
+    await toggleWatchedNote({ userId, id: params.noteId });
+    return redirect(`/notes/${params.noteId}`);
+  }
 
   return redirect("/notes");
 };
@@ -37,17 +47,26 @@ export default function NoteDetailsPage() {
   return (
     <div>
       <h3 className="text-2xl font-bold">{data.note.title}</h3>
+      <img src={data.note.url} class="max-h-52 max-w-52 float-right" />
       <p className="py-6">{data.note.body}</p>
-      <img src={data.note.url} />
       <hr className="my-4" />
       <Form method="post">
+        <input type="hidden" name="action" value="delete" />
         <button
           type="submit"
-          className="rounded bg-blue-500  py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
+          className="rounded bg-red-500 py-2 px-4 text-white hover:bg-red-600 focus:bg-red-400"
         >
           Delete
         </button>
       </Form>
-    </div>
+      <Form method="post" className="mt-4">
+        <input type="hidden" name="action" value="toggleWatched" />
+        <button
+          type="submit"
+          className="rounded bg-yellow-500 py-2 px-4 text-white hover:bg-yellow-600 focus:bg-yellow-400"
+        >
+          Toggle Watched
+        </button>
+      </Form>    </div>
   );
 }
